Perform elevated check asynchronously, use a separate function w/ switch statement
